requirements.txt
langchain==0.2.14
langchain-google-genai==1.0.7
faiss-cpu==1.13.2
pypdf==4.3.1
python-dotenv==1.0.1
tiktoken==0.7.0